changes in dtos

+++ src/main/java/br/com/xyz/zedelivery/model/dto/input/CoverageAreaInput.java
@@ -8,5 +8,5 @@ import java.util.List;
 @Getter
 public class CoverageAreaInput {
     private final String type;
-    private final List<List<List<List<Float>>>> coordinates;
+    private final List<List<List<List<Double>>>> coordinates;
 }
